requirements.txt
amqp==5.0.6
anyio==3.3.3
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.70.0
flake8==4.0.1
flower==1.0.0
greenlet==1.1.2
h11==0.12.0
humanize==3.12.0
idna==3.2
importlib-metadata==4.2.0
kombu==5.1.0
mccabe==0.6.1
prometheus-client==0.12.0
prompt-toolkit==3.0.20
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
PyMySQL==1.0.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.16.0
tornado==6.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
